requirements.txt
requests
python-telegram-bot[job-queue]>=20.0